requirements.txt
anyio==3.3.0
astroid==2.6.6
bidict==0.21.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dnspython==1.16.0
eventlet==0.31.1
Flask==2.0.1
Flask-SocketIO==5.1.1
getmac==0.8.2
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==3.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
mac-remote-command==0.0.5
MarkupSafe==2.0.1
mccabe==0.6.1
pylint==2.9.6
pyserial==3.5
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.26.0
rfc3986==1.5.0
scorebox-consoles==1.1.0
six==1.16.0
sniffio==1.2.0
toml==0.10.2
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
wrapt==1.12.1
